Move TestHelpers.hpp to unittests/test
