+++ unittests/testCollision.cpp
@@ -44,7 +44,7 @@
 #include "dart/collision/bullet/bullet.hpp"
 #include "dart/simulation/simulation.hpp"
 #include "dart/utils/utils.hpp"
-#include "TestHelpers.hpp"
+#include "test/TestHelpers.hpp"
 
 using namespace dart;
 using namespace common;
