+++ unittests/testDistance.cpp
@@ -33,7 +33,7 @@
 #include <fcl/distance.h>
 #include "dart/dart.hpp"
 #include "dart/collision/bullet/bullet.hpp"
-#include "TestHelpers.hpp"
+#include "test/TestHelpers.hpp"
 
 using namespace dart;
