+++ unittests/testJoints.cpp
@@ -32,7 +32,7 @@
 #include <array>
 #include <iostream>
 #include <gtest/gtest.h>
-#include "TestHelpers.hpp"
+#include "test/TestHelpers.hpp"
 
 #include "dart/math/Geometry.hpp"
 #include "dart/math/Helpers.hpp"
